requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.12.0
idna==3.4
lxml==4.9.2
more-itertools==9.0.0
MouseInfo==0.1.3
mss==7.0.1
multidict==6.0.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
openai==0.26.5
openai-whisper==20230124
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-pptx==0.6.21
python3-xlib==0.15
pytweening==1.0.4
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
screeninfo==0.8.1
SoundCard==0.4.2
soundfile==0.11.0
tk==0.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typing-extensions==4.4.0
urllib3==1.26.14
Wave==0.0.2
XlsxWriter==3.0.8
yarl==1.8.2
pynput==1.7.6